Creating Ships
Added a method to Main to let the user create his or her ships for the game... (computer can't make ships yet)
No placing ships method yet. Lot's of small stuff added/ edited for the greater good of the program.

util.py: added try_battlefield_int() this is the same as try_int() but doesn't convert num to an integer in the if statement